creation of simulationCA.ipynb
this notebook returning the best fit J-factor from a simulated dataset of stars and fits the degree of anisotropy in their velocity distribution, implemented by a constant beta profile (Charbonnier). To test this, the stars belong to spherical dwarf and have an isotropic velocity distribution. Therefore, the code should return beta = 0